// examples/typescript/gen/constants.ts
export enum Tags {
   Claims = 'Claims',
   Applications = 'Applications',
   Report = 'Report',
   Status = 'Status',
   Proxy = 'Proxy',
   ApplicationActions = 'Application Actions',
   ApplicationDeliverables = 'Application Deliverables',
   Profile = 'Profile',
   InternationalRecords = 'International Records',
   EducationalHistory = 'Educational History',
   CollegeSummaries = 'College Summaries',
   HighSchoolSummaries = 'High School Summaries',
   LDS = 'LDS',
   Questions = 'Questions'
}